Refactored to extract test-utils
