+++ build.gradle
@@ -72,6 +72,7 @@ subprojects {
         testCompile "org.hamcrest:hamcrest-all:1.3"
         testCompile 'junit:junit:4.12'
         testCompile "org.assertj:assertj-core:3.8.0"
+        testCompile project(":dsb-test-util")
         testCompile('org.mockito:mockito-core:1.10.19') {
             exclude group: 'org.hamcrest'// https://mvnrepository.com/artifact/net.java.dev.jna/jna
             compile group: 'net.java.dev.jna', name: 'jna', version: '4.0.0'
