+++ deploy/BlackPearlEonBrowser.html
@@ -1,0 +1,43 @@
+<html><head>
+  <SCRIPT src="http://java.com/js/dtjava.js"></SCRIPT>
+<script>
+    function launchApplication(jnlpfile) {
+        dtjava.launch(            {
+                url : 'BlackPearlEonBrowser.jnlp'
+            },
+            {
+                javafx : '8.0+'
+            },
+            {}
+        );
+        return false;
+    }
+</script>
+
+<script>
+    function javafxEmbedcom.spectralogic.dsbrowser.gui() {
+        dtjava.embed(
+            {
+                id : 'com.spectralogic.dsbrowser.gui',
+                url : 'BlackPearlEonBrowser.jnlp',
+                placeholder : 'javafx-app-placeholder',
+                width : '0',
+                height : '0'
+            },
+            {
+                javafx : '8.0+'
+            },
+            {}
+        );
+    }
+    <!-- Embed FX application into web page once page is loaded -->
+    dtjava.addOnloadCallback(javafxEmbedcom.spectralogic.dsbrowser.gui);
+</script>
+
+</head><body>
+<h2>Test page for <b>BlackPearlEonBrowser</b></h2>
+  <b>Webstart:</b> <a href='BlackPearlEonBrowser.jnlp' onclick="return launchApplication('BlackPearlEonBrowser.jnlp');">click to launch this app as webstart</a><br><hr><br>
+
+  <!-- Applet will be inserted here -->
+  <div id='javafx-app-placeholder'></div>
+</body></html>
