+++ deploy/BlackPearlEonBrowser.jnlp
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<jnlp spec="1.0" xmlns:jfx="http://javafx.com" href="BlackPearlEonBrowser.jnlp">
+  <information>
+    <title>BlackPearlEonBrowser</title>
+    <vendor>Unknown</vendor>
+    <description>BlackPearlEonBrowser</description>
+    <offline-allowed/>
+  </information>
+  <resources>
+    <j2se version="1.6+" href="http://java.sun.com/products/autodl/j2se"/>
+  </resources>
+  <jfx:javafx-desc  width="0" height="0" main-class="com.spectralogic.dsbrowser.gui.Main"  name="BlackPearlEonBrowser" />
+  <update check="background"/>
+</jnlp>
