+++ dsb-integration/build.gradle
@@ -17,6 +17,7 @@ dependencies {
     compile project(":dsb-gui")
     compile 'commons-codec:commons-codec:1.10'
     compile 'junit:junit:4.12'
+    testCompile project(":dsb-test-util")
     testCompile 'org.hamcrest:hamcrest-library:1.3'
     testCompile group: 'org.apache.commons', name: 'commons-lang3', version: '3.0'
 }
