--- dsb-integration/src/test/java/com/spectralogic/dsbrowser/integration/LoggingServiceFake.java
@@ -1,17 +1,0 @@
-package com.spectralogic.dsbrowser.integration;
-
-import com.spectralogic.dsbrowser.api.services.logging.LogType;
-import com.spectralogic.dsbrowser.api.services.logging.LoggingService;
-import io.reactivex.Observable;
-
-public class LoggingServiceFake implements LoggingService {
-    @Override
-    public Observable<LogEvent> getLoggerObservable() {
-        return Observable.fromArray();
-    }
-
-    @Override
-    public void logMessage(final String message, final LogType logType) {
-        //Just do nothing
-    }
-}
