+++ dsb-test-util/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id 'java'
+}
+
+dependencies {
+    compile project(":dsb-api")
+}
+group 'com.spectralogic.dsbrowser'
+version '5.0.0'
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+}
