+++ dsb-test-util/src/main/java/com/spectralogic/browser/gui/testUtil/LoggingServiceFake.java
@@ -1,4 +1,4 @@
-package com.spectralogic.dsbrowser.gui;
+package com.spectralogic.browser.gui.testUtil;
 
 import com.spectralogic.dsbrowser.api.services.logging.LogType;
 import com.spectralogic.dsbrowser.api.services.logging.LoggingService;
